+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 26. Added Default Fallback to Start Screen Customization.
 27. Added Old Category Purging to Start Screen Customization.
 28. Fixed Issue with Custom Start Screen Image Scanning not taking into account File Extension.
+29. Added Optional Random Theme Selection to Start Screen.
 
 ---